post-data-monitoring-chart-to-slack/requirements.txt
altair==4.1.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
entrypoints==0.3
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-bigquery==2.22.0
google-cloud-core==1.7.1
google-cloud-secret-manager==2.6.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
Jinja2==3.0.1
jsonschema==3.2.0
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.1
packaging==21.0
pandas==1.3.0
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
